Move getConnectionName to wallet api
