+++ src/components/Web3Provider/index.tsx
@@ -1,7 +1,7 @@
 import { Web3ReactHooks, Web3ReactProvider } from '@web3-react/core'
 import { Connector } from '@web3-react/types'
 import { Connection } from '@cow/modules/wallet/web3-react/utils/connection/connections'
-import { getConnectionName } from '@cow/modules/wallet/web3-react/utils/connection'
+import { getConnectionName } from '@cow/modules/wallet/api/utils/connections'
 import useEagerlyConnect from 'hooks/useEagerlyConnect'
 import useOrderedConnections from 'hooks/useOrderedConnections'
 import { ReactNode, useMemo } from 'react'
