+++ src/cow-react/modules/wallet/api/utils/connections.ts
@@ -1,0 +1,18 @@
+import { ConnectionType } from "../types"
+
+export function getConnectionName(connectionType: ConnectionType, isMetaMask?: boolean) {
+  switch (connectionType) {
+    case ConnectionType.INJECTED:
+      return isMetaMask ? 'MetaMask' : 'Injected'
+    case ConnectionType.COINBASE_WALLET:
+      return 'Coinbase Wallet'
+    case ConnectionType.WALLET_CONNECT:
+      return 'WalletConnect'
+    case ConnectionType.FORTMATIC:
+      return 'Fortmatic'
+    case ConnectionType.NETWORK:
+      return 'Network'
+    case ConnectionType.GNOSIS_SAFE:
+      return 'Gnosis Safe'
+  }
+}
