+++ src/cow-react/modules/wallet/web3-react/containers/ConnectWalletOptions/CoinbaseWalletOption.tsx
@@ -1,6 +1,6 @@
 import { ConnectionType } from '@cow/modules/wallet'
 import { coinbaseWalletConnection } from '@cow/modules/wallet/web3-react/utils/connection/connections'
-import { getConnectionName } from '@cow/modules/wallet/web3-react/utils/connection'
+import { getConnectionName } from '@cow/modules/wallet/api/utils/connections'
 
 import { useSelectedWallet } from 'state/user/hooks'
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet' // MOD
