+++ src/cow-react/modules/wallet/web3-react/containers/ConnectWalletOptions/FortmaticOption.tsx
@@ -1,7 +1,7 @@
 import { ConnectionType } from '@cow/modules/wallet'
 
 import { fortmaticConnection } from '@cow/modules/wallet/web3-react/utils/connection/connections'
-import { getConnectionName } from '@cow/modules/wallet/web3-react/utils/connection'
+import { getConnectionName } from '@cow/modules/wallet/api/utils/connections'
 
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet'
 import { ConnectWalletOption } from '@cow/modules/wallet/api/pure/ConnectWalletOption'
