+++ src/cow-react/modules/wallet/web3-react/utils/connection/index.ts
@@ -55,19 +55,3 @@ export function getConnection(c: Connector | ConnectionType): Connection {
   }
 }
 
-export function getConnectionName(connectionType: ConnectionType, isMetaMask?: boolean) {
-  switch (connectionType) {
-    case ConnectionType.INJECTED:
-      return isMetaMask ? 'MetaMask' : 'Injected'
-    case ConnectionType.COINBASE_WALLET:
-      return 'Coinbase Wallet'
-    case ConnectionType.WALLET_CONNECT:
-      return 'WalletConnect'
-    case ConnectionType.FORTMATIC:
-      return 'Fortmatic'
-    case ConnectionType.NETWORK:
-      return 'Network'
-    case ConnectionType.GNOSIS_SAFE:
-      return 'Gnosis Safe'
-  }
-}
